+++ sdk/batch/azure-mgmt-batch/azure/mgmt/batch/aio/_batch_management_client.py
@@ -8,9 +8,12 @@
 
 from copy import deepcopy
 from typing import Any, Awaitable, TYPE_CHECKING
+from typing_extensions import Self
 
+from azure.core.pipeline import policies
 from azure.core.rest import AsyncHttpResponse, HttpRequest
 from azure.mgmt.core import AsyncARMPipelineClient
+from azure.mgmt.core.policies import AsyncARMAutoResourceProviderRegistrationPolicy
 
 from .. import models as _models
 from .._serialization import Deserializer, Serializer
@@ -21,6 +24,7 @@ from .operations import (
     BatchAccountOperations,
     CertificateOperations,
     LocationOperations,
+    NetworkSecurityPerimeterOperations,
     Operations,
     PoolOperations,
     PrivateEndpointConnectionOperations,
@@ -54,6 +58,9 @@ class BatchManagementClient:  # pylint: disable=client-accepts-api-version-keywo
      azure.mgmt.batch.aio.operations.PrivateEndpointConnectionOperations
     :ivar pool: PoolOperations operations
     :vartype pool: azure.mgmt.batch.aio.operations.PoolOperations
+    :ivar network_security_perimeter: NetworkSecurityPerimeterOperations operations
+    :vartype network_security_perimeter:
+     azure.mgmt.batch.aio.operations.NetworkSecurityPerimeterOperations
     :param credential: Credential needed for the client to connect to Azure. Required.
     :type credential: ~azure.core.credentials_async.AsyncTokenCredential
     :param subscription_id: The Azure subscription ID. This is a GUID-formatted string (e.g.
@@ -61,7 +68,7 @@ class BatchManagementClient:  # pylint: disable=client-accepts-api-version-keywo
     :type subscription_id: str
     :param base_url: Service URL. Default value is "https://management.azure.com".
     :type base_url: str
-    :keyword api_version: Api Version. Default value is "2024-02-01". Note that overriding this
+    :keyword api_version: Api Version. Default value is "2024-07-01". Note that overriding this
      default value may result in unsupported behavior.
     :paramtype api_version: str
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
@@ -78,7 +85,25 @@ class BatchManagementClient:  # pylint: disable=client-accepts-api-version-keywo
         self._config = BatchManagementClientConfiguration(
             credential=credential, subscription_id=subscription_id, **kwargs
         )
-        self._client: AsyncARMPipelineClient = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
+        _policies = kwargs.pop("policies", None)
+        if _policies is None:
+            _policies = [
+                policies.RequestIdPolicy(**kwargs),
+                self._config.headers_policy,
+                self._config.user_agent_policy,
+                self._config.proxy_policy,
+                policies.ContentDecodePolicy(**kwargs),
+                AsyncARMAutoResourceProviderRegistrationPolicy(),
+                self._config.redirect_policy,
+                self._config.retry_policy,
+                self._config.authentication_policy,
+                self._config.custom_hook_policy,
+                self._config.logging_policy,
+                policies.DistributedTracingPolicy(**kwargs),
+                policies.SensitiveHeaderCleanupPolicy(**kwargs) if self._config.redirect_policy else None,
+                self._config.http_logging_policy,
+            ]
+        self._client: AsyncARMPipelineClient = AsyncARMPipelineClient(base_url=base_url, policies=_policies, **kwargs)
 
         client_models = {k: v for k, v in _models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
@@ -99,8 +124,13 @@ class BatchManagementClient:  # pylint: disable=client-accepts-api-version-keywo
             self._client, self._config, self._serialize, self._deserialize
         )
         self.pool = PoolOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.network_security_perimeter = NetworkSecurityPerimeterOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
 
-    def _send_request(self, request: HttpRequest, **kwargs: Any) -> Awaitable[AsyncHttpResponse]:
+    def _send_request(
+        self, request: HttpRequest, *, stream: bool = False, **kwargs: Any
+    ) -> Awaitable[AsyncHttpResponse]:
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
@@ -120,12 +150,12 @@ class BatchManagementClient:  # pylint: disable=client-accepts-api-version-keywo
 
         request_copy = deepcopy(request)
         request_copy.url = self._client.format_url(request_copy.url)
-        return self._client.send_request(request_copy, **kwargs)
+        return self._client.send_request(request_copy, stream=stream, **kwargs)  # type: ignore
 
     async def close(self) -> None:
         await self._client.close()
 
-    async def __aenter__(self) -> "BatchManagementClient":
+    async def __aenter__(self) -> Self:
         await self._client.__aenter__()
         return self